server/ai_service/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
scipy==1.15.1
matplotlib==3.10.0
seaborn==0.13.2
